fix/refactor: The importening 2 - importing boogaloo
* Fixed a lot of imports in the view model classes.

* The base ViewModel class mustn't import the view model collection, or else cause a circular dependency, which results in import errors. Instead, the input and button components must be handled by the actor-sheet and item-sheet, explicitly.
