+++ module/sheets/item-sheet.mjs
@@ -1,10 +1,10 @@
-import ViewModelCollection from "../utils/viewmodel-collection.mjs";
 import AmbersteelBaseItemSheet from "./subtypes/item/ambersteel-base-item-sheet.mjs";
 import AmbersteelSkillItemSheet from "./subtypes/item/ambersteel-skill-item-sheet.mjs";
 import AmbersteelFateCardItemSheet from "./subtypes/item/ambersteel-fate-item-sheet.mjs";
 import AmbersteelInjuryItemSheet from "./subtypes/item/ambersteel-injury-item-sheet.mjs";
 import AmbersteelIllnessItemSheet from "./subtypes/item/ambersteel-illness-item-sheet.mjs";
 import * as SheetUtil from "../utils/sheet-utility.mjs";
+import ViewModelCollection from "../utils/viewmodel-collection.mjs";
 
 export class AmbersteelItemSheet extends ItemSheet {
   /**
@@ -73,6 +73,13 @@ export class AmbersteelItemSheet extends ItemSheet {
    */
   get viewModel() { return this._viewModel; }
 
+  /**
+   * @private
+   * @todo Remove and properly integrate inputs and buttons into the "new" view model system. 
+   * @type {ViewModelCollection}
+   */
+  _inputsAndButtons = undefined;
+
   /** 
    * Returns an object that represents sheet and enriched item data. 
    * 
@@ -88,8 +95,9 @@ export class AmbersteelItemSheet extends ItemSheet {
     SheetUtil.enrichData(context);
 
     // Prepare view model. 
-    this.viewModel = this.subType.getViewModel(context);
-    this.viewModel.applyViewState(this._getViewState(this.item.id));
+    this._viewModel = this.subType.getViewModel(context);
+    this._viewModel.applyViewState(this._getViewState(this.item.id));
+    this._inputsAndButtons = new ViewModelCollection();
 
     this.subType.prepareDerivedData(context);
 
@@ -131,6 +139,7 @@ export class AmbersteelItemSheet extends ItemSheet {
 
     // Activate view model bound event listeners. 
     this.viewModel.activateListeners(html, isOwner, isEditable);
+    this._inputsAndButtons.activateListeners(html, isOwner, isEditable);
 
     // -------------------------------------------------------------
     if (!isOwner) return;
@@ -145,6 +154,7 @@ export class AmbersteelItemSheet extends ItemSheet {
   async close() {
     this._setViewState(this.item.id);
     this.viewModel.dispose();
+    this._inputsAndButtons.dispose();
 
     return super.close();
   }
