+++ templates/actor/parts/actor-assets-viewmodel.mjs
@@ -1,6 +1,5 @@
-import { TEMPLATES } from "../../templatePreloader.mjs";
-import ViewModel from "../../module/components/viewmodel.mjs";
-import SheetViewModel from "../sheet-viewmodel.mjs";
+import { TEMPLATES } from "../../../module/templatePreloader.mjs";
+import SheetViewModel from "../../sheet-viewmodel.mjs";
 
 export default class ActorAssetsViewModel extends SheetViewModel {
   /** @override */
