+++ templates/actor/parts/actor-beliefs-fate-viewmodel.mjs
@@ -1,6 +1,5 @@
-import { TEMPLATES } from "../../templatePreloader.mjs";
-import ViewModel from "../../module/components/viewmodel.mjs";
-import SheetViewModel from "../sheet-viewmodel.mjs";
+import { TEMPLATES } from "../../../module/templatePreloader.mjs";
+import SheetViewModel from "../../sheet-viewmodel.mjs";
 import ActorBeliefsViewModel from "./actor-beliefs-viewmodel.mjs";
 import ActorFateViewModel from "./actor-fate-viewmodel.mjs";
