+++ templates/actor/parts/actor-personals-viewmodel.mjs
@@ -1,6 +1,5 @@
-import ViewModel from "../../module/components/viewmodel.mjs";
-import { TEMPLATES } from "../../templatePreloader.mjs";
-import SheetViewModel from "../sheet-viewmodel.mjs";
+import { TEMPLATES } from "../../../module/templatePreloader.mjs";
+import SheetViewModel from "../../sheet-viewmodel.mjs";
 
 export default class ActorPersonalsViewModel extends SheetViewModel {
   /** @override */
