+++ templates/item/skill-ability/skill-ability-table-viewmodel.mjs
@@ -1,5 +1,5 @@
-import ViewModel from "../../module/components/viewmodel.mjs";
-import SheetViewModel from "../sheet-viewmodel.mjs";
+import { TEMPLATES } from "../../../module/templatePreloader.mjs";
+import SheetViewModel from "../../sheet-viewmodel.mjs";
 import SkillAbilityListItemViewModel from "./skill-ability-list-item-viewmodel.mjs";
 
 export default class SkillAbilityTableViewModel extends SheetViewModel {
